mlops/pyproject.toml
[project]
name = "hotel-booking"
dynamic = ["version"]
description = "Code accompanying MLOps with Databrick O'Reilly book"
requires-python = ">=3.12, <3.13"
dependencies = [
    "mlflow==3.8.1",
    "cffi==1.17.1",
    "cloudpickle==3.1.1",
    "numpy==2.4.0",
    "pandas==2.3.0",
    "pyarrow==22.0.0",
    "databricks-sdk==0.85.0",
    "scikit-learn==1.7.0",
    "lightgbm==4.6.0",
    "scipy==1.16.0",
    "pydantic==2.11.7",
    "loguru==0.7.3",
    "python-dotenv==1.1.1",
    "ray[tune]==2.52.1",
    "optuna==4.4.0",
    "sdv==1.24.0",
    "databricks-feature-engineering==0.14.0",
    "databricks-feature-lookup==1.11.0"
]

[project.optional-dependencies]
dev = ["databricks-connect>=17.3.2, <18",
       "ipykernel>=6.29.5, <7",
       "pip>=25.0.1, <26",
       "pre-commit>=4.1.0, <5",
       "pytest>=8.3.4, <9"]

[build-system]
requires = [
  "setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where =["src"]
include = ["hotel_booking*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
addopts = "-s --no-header --no-summary"
testpaths = [ "tests" ]
pythonpath = [ "." ]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = [
    "F",    # pyflakes rules
    "E",    # pycodestyle error rules
    "W",    # pycodestyle warning rules
    "B",    # flake8-bugbear rules
    "I",    # isort rules
    "UP",   # pyupgrade rules
    "SIM",  # flake8-simplify rules
    "ERA",  # eradicate rules
    "C",    # pylint convention rules
    "ANN"   # flake8-annotations rules
]
ignore = [
    "D107", # Missing docstring in `__init__`
    "D203", # incorrect-blank-line-before-class
    "D213", # multi-line-summary-second-line
    "C901", # too complex
    "ERA001", # commented out code
]

[tool.ruff.format]
indent-style = "space"
